Fix `Localidade` pk type to `BIGSERIAL`.
Due to @fccoelho comment https://github.com/fccoelho/AlertaDengueCaptura/pull/8#issuecomment-146313054

+++ dbschema/Banco_de_dados_do_Alerta_Dengue_create.sql
@@ -136,7 +136,7 @@ CREATE TABLE "Municipio"."Localidade" (
     "nome" varchar(32)  NOT NULL,
     "populacao" int  NOT NULL,
     "geojson" text  NOT NULL,
-    "id" int  NOT NULL,
+    "id" BIGSERIAL NOT NULL,
     "Municipio_geocodigo" int  NOT NULL,
     CONSTRAINT "Localidade_pk" PRIMARY KEY ("id")
 );
